backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
websockets
python-dotenv==1.0.0
pyaudio==0.2.14
google-genai
scrolledtext